chore: move web service to subfolder
