+++ .gitignore
@@ -88,4 +88,4 @@ build/
 .nosync
 
 # Web Validator Authentication Credential
-web-api/src/main/resources/web-based-gtfs-validator-3c339fa28589.json
+web/service/src/main/resources/web-based-gtfs-validator-3c339fa28589.json
