+++ web/service/src/main/java/org/mobilitydata/gtfsvalidator/web/service/UploadFileResponse.java
@@ -1,4 +1,4 @@
-package org.mobilitydata.gtfsvalidator.webapi;
+package org.mobilitydata.gtfsvalidator.web.service;
 
 public class UploadFileResponse {
